+++ tests/conftest.py
@@ -148,8 +148,9 @@ def test_df_year():
 # minimal IamDataFrame for specifically testing 'time'-column features
 @pytest.fixture(scope="function")
 def test_df_time():
-    tdf = TEST_DF.rename({2005: TEST_DTS[0], 2010: TEST_DTS[1]}, axis="columns")
-    df = IamDataFrame(data=tdf)
+    df = IamDataFrame(
+        data=TEST_DF.rename({2005: TEST_DTS[0], 2010: TEST_DTS[1]}, axis="columns")
+    )
     for i in META_COLS:
         df.set_meta(META_DF[i])
     yield df